requirements.txt
FastAPI
uvicorn[standard]
python-dotenv
openai
python-multipart
pymongo[srv]